+++ test/distributed_ranges/common/iota.cpp
@@ -13,7 +13,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "xhp-tests.hpp"
+#include "xhp_tests.hpp"
 
 // Fixture
 template <typename DistVecT> class IotaTest : public testing::Test {
