+++ test/distributed_ranges/common/sycl_utils.cpp
@@ -13,7 +13,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "xhp-tests.hpp"
+#include "xhp_tests.hpp"
 
 #ifdef SYCL_LANGUAGE_VERSION
