+++ test/distributed_ranges/shp/copy.hpp
@@ -15,7 +15,7 @@
 
 #pragma once
 
-#include "xhp-tests.hpp"
+#include "xhp_tests.hpp"
 
 template <typename AllocT> class CopyTest : public testing::Test {
 public:
