+++ test/distributed_ranges/shp/transform.cpp
@@ -13,7 +13,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "xhp-tests.hpp"
+#include "xhp_tests.hpp"
 
 template <typename AllocT> class TransformTest : public testing::Test {
 public:
